requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
qrcode~=7.4.2